docs: Document our loadingTriangle component
